move search interfaces into a separate directory
