+++ types/api/search/index.d.ts
@@ -1,0 +1,5 @@
+import { Hotel } from "./hotel";
+
+export namespace Search {
+  export { Hotel };
+}
